Rename math.py to layer_math.py
* Fix #903
