+++ python/paddle/trainer_config_helpers/__init__.py
@@ -20,6 +20,4 @@ from layers import *
 from networks import *
 from optimizers import *
 from attrs import *
-
-# This will enable operator overload for LayerOutput
-import math as layer_math
+import layer_math
